@@ -11,7 +11,7 @@
     "server-d": "webpack-dev-server --inline --hot",
     "server-pm2": "pm2 start server.js --name uops",
     "server-pm2-restart": "pm2 restart server.js",
-    "server-p": "node server.js"
+    "server-p": "babel-node server.js"
   },
   "repository": {
     "type": "git",
@@ -24,21 +24,18 @@
   },
   "homepage": "https://github.com/yangblgaoyc/reacttest#readme",
   "devDependencies": {
-    "@babel/cli": "^7.2.0",
-    "@babel/core": "^7.2.0",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.2.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-stage-0": "^7.0.0",
-    "@babel/preset-stage-3": "^7.0.0",
-    "@babel/register": "^7.0.0",
     "autoprefixer": "^9.4.2",
+    "babel-core": "^6.26.3",
     "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.4",
+    "babel-loader": "^7.1.5",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-3": "^6.24.1",
     "clean-webpack-plugin": "^1.0.0",
     "css-loader": "^1.0.1",
-    "eslint": "^5.10.0",
+    "eslint": "^5.12.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-friendly-formatter": "^4.0.1",
     "eslint-loader": "^2.1.1",
